+++ tests/Feature/ConfigTest.php
@@ -1,6 +1,8 @@
 <?php
 
-namespace Sfneal\PostOffice\Tests;
+namespace Sfneal\PostOffice\Tests\Feature;
+
+use Sfneal\PostOffice\Tests\TestCase;
 
 class ConfigTest extends TestCase
 {
