+++ tests/Feature/ViewsTest.php
@@ -1,8 +1,9 @@
 <?php
 
-namespace Sfneal\PostOffice\Tests;
+namespace Sfneal\PostOffice\Tests\Feature;
 
 use Sfneal\PostOffice\Tests\Assets\InvoiceUnpaidMailable;
+use Sfneal\PostOffice\Tests\TestCase;
 use Sfneal\Users\Models\User;
 
 class ViewsTest extends TestCase
